Config scripts updated to use env file variables

+++ 19_docker_compose_init.sh
@@ -3,25 +3,31 @@
 # Initial config of Docker Compose
 #
 
-STACKDIR=${HOME}/smarthomestack
+echo "=> Docker Compose init started"
 
+if [ ! -e .env ]
+then
+    echo "<!> Please create proper .env file! Exiting."
+    exit 1
+fi
+
+set -o allexport
+source .env
+set +o allexport
+
+# ${STACKDIR} is from .env
 cd ${STACKDIR}
-#if [ ! -e pgsql.env ]
-#then
-#    cp pgsql.env-example pgsql.env
-#    echo "Don't forget to update pgsql.env file!"
-#fi
 
 if [ ! -e influxdb.env ]
 then
     cp influxdb.env-example influxdb.env
-    echo "Don't forget to update influxdb.env file!"
+    echo "=> Don't forget to update influxdb.env file!"
 fi
 
 if [ ! -e mqtt.env ]
 then
     cp mqtt.env-example mqtt.env
-    echo "Don't forget to update mqtt.env file!"
+    echo "=> Don't forget to update mqtt.env file!"
 fi
 
 COMPOSEFILE=${STACKDIR}/docker-compose.yml
@@ -29,17 +35,19 @@ echo "docker clean-up"
 docker system prune -f;docker image prune -f;docker volume prune -f
 
 echo "Validating: ${COMPOSEFILE}"
-docker-compose -f ${COMPOSEFILE} config
+${COMPOSECOMMAND} -f ${COMPOSEFILE} config
 read -p "=> Validating finished, press any key"
 
 echo "Initiate downloading of docker images from Docker Hub"
-docker-compose -f ${COMPOSEFILE} pull
+${COMPOSECOMMAND} -f ${COMPOSEFILE} pull
 
-echo "First docker-compose run, errors messages expected due to lack of proper config files"
-docker-compose -f ${COMPOSEFILE} up -d
+echo "First ${COMPOSECOMMAND} run, errors messages expected due to lack of proper config files"
+${COMPOSECOMMAND} -f ${COMPOSEFILE} up -d
 
 echo "Stopping all containers"
-docker-compose -f ${COMPOSEFILE} down
+${COMPOSECOMMAND} -f ${COMPOSEFILE} down
 
 echo "Docker clean-up again"
 docker system prune -f;docker image prune -f;docker volume prune -f
+
+echo "=> Docker Compose init completed"
